refactor data structures, operations are extending the same operation now.
