+++ state-engine/src/main/java/scheduler/context/og/OGSAContext.java
@@ -1,6 +1,6 @@
 package scheduler.context.og;
 
-public class OGDFSAContext extends OGDFSContext {
+public class OGDFSAContext extends OGSContext {
 
     public int rollbackLevel = -1; // initialized to 0 if thread not required to be rollbacked.
     public boolean isRollbacked = false; // initialized to 0 if thread not required to be rollbacked.
