+++ state-engine/src/main/java/scheduler/struct/AbstractOperation.java
